backend/requirements.txt
fastapi>=0.115.0,<1.0
pydantic~=2.6.4
uvicorn>=0.30.6,<1.0
python-multipart>=0.0.10,<1.0
httpx>=0.27.2, <1.0
PyJWT[crypto]>=2.9.0,<3.0
https://github.com/nhairs/python-json-logger/releases/download/v3.1.0/python_json_logger-3.1.0-py3-none-any.whl#egg=python-json-logger
